ARM64: configs: Configure optimal minimum CPU frequencies for sm8150
These frequencies have been selected by weighing absolute power usage
vs. efficiency (mW/perf -- the lower the better) for all the frequencies
available on each cluster.

Performance was measured by running an in-kernel RC4 benchmark
simultaneously on all the cores in a cluster (minus one for
housekeeping, except Prime where the first core of the little cluster
was used for housekeeping). Power usage was measured while the benchmark
was running using the integrated PMIC (PM8150B v2.0).

768000 kHz, 1056000 kHz, and 1171200 kHz have each been selected by
looking for the most efficient frequency with a relatively low power
usage. That is, when mW/perf and power usage (mW) are graphed together,
the ideal frequency is one with very low mW/perf and mW, such that the
lines are intersecting. Since this is obviously not possible in
practice, we look for the frequency closest to matching this
description.

We aren't using the *most* efficient freqs because PM QoS abuse and high
overall CPU usage causes higher drain on them by reducing deep idle
time and utilizing the CPU more, which makes the frequency's high
absolute power usage show.

As a result of this careful selection, the new frequencies should
benefit both performance *and* battery endurance in real-world
scenarios.

Full data with graphs: https://docs.google.com/spreadsheets/d/1gOG4MVDzOvfAtXzifL0EQm-BPHdGkZ9SHThybRpm6R8/edit

Thanks to Sultan Alsawaf (kerneltoast) for creating the benchmark used
in this test.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>