Revert "msm: vidc: Inline HAL_FLIP value with omx counterpart"
This break screenshot taking, by generating an image with
the wrong metadata.

The commit mentions that this patch inlines the available
image flip values with the OMX component, but the only
place I could find similar values defined are display HAL,
and they followed the old values.

enum struct DRMRotation {
  FLIP_H = 0x1,
  FLIP_V = 0x2,
  ROT_180 = FLIP_H | FLIP_V,
  ROT_90 = 0x4,
};

Bring back the old values to fix this.

This reverts commit 73643c26eb7f10dc878a2a5b3962e0373029da5f.

Change-Id: I859119c96b316a64b3e3bd35b5b5319d27e64a84
Signed-off-by: Rajan Palaniya <razorpaul@gmail.com>